Move genesis run commands implementation into era-based
