+++ cardano-cli/src/Cardano/CLI/Legacy/Run.hs
@@ -4,9 +4,9 @@ module Cardano.CLI.Legacy.Run
   ( runLegacyCmds
   ) where
 
+import           Cardano.CLI.EraBased.Run.Genesis
 import           Cardano.CLI.Legacy.Options
 import           Cardano.CLI.Legacy.Run.Address
-import           Cardano.CLI.Legacy.Run.Genesis
 import           Cardano.CLI.Legacy.Run.Governance
 import           Cardano.CLI.Legacy.Run.Key
 import           Cardano.CLI.Legacy.Run.Node
